Elevated: Use more specific types since gcc and msvc disagree

+++ games/elevated/stats/PowerStatsListener.cpp
@@ -13,9 +13,9 @@ void PowerStatsListener::on_elevator_moved(Time, Height height_travelled, Height
     m_distance_moved[state.filled_capacity()] += height_travelled;
 }
 
-size_t PowerStatsListener::total_distance_travelled() const
+uint64_t PowerStatsListener::total_distance_travelled() const
 {
-    return std::accumulate(m_distance_moved.begin(), m_distance_moved.end(), size_t(0), [](uint64_t acc, auto& entry) {
+    return std::accumulate(m_distance_moved.begin(), m_distance_moved.end(), uint64_t(0), [](uint64_t acc, auto& entry) {
         return acc + entry.second;
     });
 }
